Move App to its own file
